third_party/blink/web_tests/external/wpt/dom/nodes/adoption.window-expected.txt

This is a testharness.js-based test.
[FAIL] adoptNode() and DocumentFragment with host
  assert_equals: expected Document node with 0 children but got Document node with 2 children
Harness: the test ran to completion.
